Numerous QOL Changes
* exampleSite/config.toml - changed default theme directory to 
hugo-uilite
opposed to uilite.
  * The user won't have to rename the
theme in /themes or change config.toml to 'uilite' when running an
example site.

* Added support for awesome fas icons (opposed to only fab).
This is useful if a user wants to have a mail icon for example.
**by default this creates a new envelope icon in the exampleSite
sidebar** can be removed in exampleSite/data/socialfas.json.
  * New fas icons should be placed in exampleSite/data/socialfas.json.
  * Fas icons are rendered after the fab icons.

* Added period after description in exampleSite/data/sidebar.json
* Added a newline after variable declaration in
layouts/partials/contact.html and layouts/partials/sidebar.html to match
with other html files.